Changes for flowserv 0.9.0
